[boundary] Fix error rb.validator.ValidateCreateRectangle undefined (type *validator.Validator is pointer to interface, not interface.

+++ APP/boundary/io/rectangle_input.go
@@ -48,8 +48,8 @@ type GetRectangleInput struct {
 type UpdateRectangleInput struct {
 	ID       string
 	Name     string
-	Width    int
-	Height   int
+	Width    float64
+	Height   float64
 	Position common.Position
 	Color    common.Color
 }
